+++ spike/strawberry_poc/models/rupture_set.py
@@ -18,7 +18,7 @@ from data.dynamo import create_file, get_file, get_thing, list_files
 from data.models import RuptureSetData
 
 from .automation_task import RuptureGenerationTask  # noqa: F401 — re-exported for schema.py
-from .common import KeyValuePair, KeyValuePairInput
+from .common import BigInt, KeyValuePair, KeyValuePairInput
 from .file_interface import FileInterface
 
 # ── RuptureSet ────────────────────────────────────────────────────────────────
@@ -73,7 +73,7 @@ class CreateRuptureSetInput:
     file_name: str
     produced_by: strawberry.ID
     md5_digest: str | None = None
-    file_size: int | None = None
+    file_size: BigInt | None = None
     created: str | None = None
     fault_models: list[str] | None = None
     metrics: list[KeyValuePairInput] | None = None
